requirements.txt
quart
requests
tapo
python-dotenv